requirements.txt
Django==6.0.1
pyserial